// src/redux/actions/types/filter.ts
enum FilterTypes {
  SET_FILTER_DATA = 'Filter/SET_FILTER_DATA',
  FETCH_BU = 'Filter/FETCH_BU',
  FETCH_TEAMS = 'Filter/FETCH_TEAMS',
  FETCH_AGENTS = 'Filter/FETCH_AGENTS',
  SEARCH_TEAMS = 'Filter/SEARCH_TEAMS',
  SEARCH_AGENTS = 'Filter/SEARCH_AGENTS',
  SET_SEARCH_DATA = 'Filter/SET_SEARCH_DATA',
  SET_SEARCH_ACTIVE = 'Filter/SET_SEARCH_ACTIVE',
  SET_CHECKED_ITEMS = 'Filter/SET_CHECKED_ITEMS',
  SET_INIT_CHECKED_ITEMS = 'Filter/SET_INIT_CHECKED_ITEMS',
  CHANGE_SIDEBAR = 'Filter/CHANGE_SIDEBAR',
  MERGE_DATA = 'Filter/MERGE_DATA',
  RESTORE_FILTER_SETTINGS = 'Filter/RESTORE_FILTER_SETTINGS',
  SHOW_EMPTY = 'Filter/SHOW_EMPTY',
}
export default FilterTypes;